pyproject.toml
[project]
name = "opensarch-demo"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "awscurl>=0.36",
    "boto3>=1.42.47",
    "fastapi>=0.128.0",
    "jinja2>=3.1.6",
    "opensearch-py>=3.1.0",
    "python-dotenv>=1.2.1",
    "requests-aws4auth>=1.3.1",
    "scylla-driver>=3.29.7",
    "sentence-transformers>=5.2.2",
    "uvicorn>=0.40.0",
]
